requirements.txt
# General
wheel
torch
pandas
scikit-learn
simpletransformers
fastapi[all]
datasets

# Ontology
RdfLib

urllib3<2
